+++ client/src/App.css
@@ -1,135 +1,1 @@
-/* Colors */
-/*
-dark blue: #476A6F
-light blue: #258EA6
-dark green: #519E8A
-medium green: #7EB09B
-light green: #A3C4BC
-*/
 
-.wrapper {
-  height: 100vh;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-color: #258ea6;
-}
-
-.form-wrapper {
-  width: 400px;
-  display: flex;
-  flex-direction: column;
-  padding: 20px 40px;
-  border-radius: 10px;
-  box-shadow: 0px 10px 50px #555;
-  background-color: #ffffff;
-}
-
-form {
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-h1 {
-  text-align: center;
-  width: 100%;
-  color: #111;
-  font-weight: lighter;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
-    Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-}
-
-label {
-  font-size: 0.8em;
-  margin-bottom: 0.25em;
-  color: #222;
-  font-weight: lighter;
-}
-
-input {
-  padding: 10px 10px;
-  border-radius: 5px;
-  outline: none;
-  border: 1px solid #cfcfcf;
-}
-
-input::placeholder {
-  font-size: 1.2em;
-  font-weight: lighter;
-  color: #999;
-}
-
-input.error {
-  border: 1px solid red;
-}
-
-.errorMessage {
-  color: red;
-  font-size: 0.75em;
-  display: relative;
-}
-
-.firstName {
-  margin-right: 1%;
-}
-
-.lastName {
-  margin-left: 1%;
-}
-
-.firstName,
-.lastName,
-.email,
-.password {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 15px;
-}
-
-.firstName,
-.lastName {
-  width: 49%;
-}
-
-.email,
-.password {
-  width: 100%;
-}
-
-.createAccount {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.createAccount button {
-  background-color: #519e8a;
-  color: #fff;
-  border: 2px solid #fff;
-  width: 100%;
-  margin-top: 1em;
-  padding: 8px 0px;
-  font-size: 1em;
-  font-weight: lighter;
-  letter-spacing: 1px;
-  margin-bottom: 0.25em;
-}
-
-.createAccount button:hover {
-  color: #519e8a;
-  background-color: #fff;
-  border: 2px solid #519e8a;
-}
-
-.createAccount small {
-  color: #999;
-  font-weight: lighter;
-}
-
-.headerTitle {
-  font-weight: bold;
-}
