+++ client/src/pages/NoMatch.js
@@ -1,0 +1,19 @@
+import React from "react";
+import { Col, Row, Container } from "../components/Grid";
+import Jumbotron from "../components/Jumbotron";
+
+function NoMatch() {
+  return (
+    <Container fluid>
+      <Row>
+        <Col size="md-12">
+          <Jumbotron>
+            <h1>404 Page Not Found</h1>
+          </Jumbotron>
+        </Col>
+      </Row>
+    </Container>
+  );
+}
+
+export default NoMatch;
